[PROGRESSION] Added 3 Cards @ DataManagement Tab with Static Required Inputs
INFO: Those following textfield is only static. Thats because, we need more time to implement dynamic bindings by making user add more what is required and what is not. The following textfields are just not placeholders but that is based on the templates of the inventory system that they have...
Deducing the count of inputs is possible, upon by the wish of the client.

+ Replaced Dialogs into Default One from Active KivyMD Repo
+ Added Prototype homepage for helping people
+ Added Potential Clock Max Iterations to 30