[DEVELOPMENT] Finalized Import Binding and Export Pull Out Design
Current Goal for Designs:
+ Data Status Report Design
+ Dashboard Design (Make this one minimal)

Current Goal for Functions:
EVERYTHING IS NOT YET FINISHED INRODER TO DO THIS

Database is needed!
Excel Dynamic Path is also needed!
Account Section Functions is not there as well!

TL;DR -> Most of the designs is finished, clarification on functions is not yet started on building... It should take one day in attempt to do designs on two more sections... That could be for now, break time on coding!